+++ src/components/ColorPicker/__tests__/ColorPicker.interactions.js
@@ -1,5 +1,5 @@
 import { userEvent, within } from "@storybook/testing-library";
-import { contentColorsByName } from "../../../utils/colors-vars-map";
+import { ContentColorByName } from "../../../utils/colors-vars-map";
 import { resetFocus } from "../../../__tests__/interactions-helper";
 import { ComponentDefaultTestId } from "../../../tests/constants";
 import { typeMultipleTimes, interactionSuite } from "../../../tests/interactions-utils";
@@ -7,24 +7,24 @@ import { expect } from "@storybook/jest";
 import { getTestId } from "../../../tests/test-ids-utils";
 
 async function selectAndResetWithKeyboard(canvas) {
-  await clickOnColor(canvas, contentColorsByName.BRIGHT_GREEN);
-  await expectColorToBeSelected(canvas, contentColorsByName.BRIGHT_GREEN);
+  await clickOnColor(canvas, ContentColorByName.BRIGHT_GREEN);
+  await expectColorToBeSelected(canvas, ContentColorByName.BRIGHT_GREEN);
 
   //move to a color in the last row
   await typeMultipleTimes("{arrowDown}", 7);
-  await expectColorToBeActive(canvas, contentColorsByName.STEEL);
+  await expectColorToBeActive(canvas, ContentColorByName.STEEL);
 
   //move to a color in the last row
   await userEvent.keyboard(" ");
-  await expectColorToBeActive(canvas, contentColorsByName.STEEL);
-  await expectColorToBeSelected(canvas, contentColorsByName.STEEL);
-  await expectColorToBeNotSelected(canvas, contentColorsByName.BRIGHT_GREEN);
+  await expectColorToBeActive(canvas, ContentColorByName.STEEL);
+  await expectColorToBeSelected(canvas, ContentColorByName.STEEL);
+  await expectColorToBeNotSelected(canvas, ContentColorByName.BRIGHT_GREEN);
 
   //move to the "Clear color" button
   await userEvent.keyboard("{arrowDown}");
-  await expectColorToBeNotActive(canvas, contentColorsByName.STEEL);
+  await expectColorToBeNotActive(canvas, ContentColorByName.STEEL);
   await userEvent.keyboard("{Enter}");
-  await expectColorToBeNotSelected(canvas, contentColorsByName.STEEL);
+  await expectColorToBeNotSelected(canvas, ContentColorByName.STEEL);
 }
 
 export const noColorInteractionSuite = interactionSuite({
@@ -35,27 +35,27 @@ export const noColorInteractionSuite = interactionSuite({
 });
 
 async function selectMultiColorsWithKeyboardAndMouse(canvas) {
-  await clickOnColor(canvas, contentColorsByName.DARK_PURPLE);
-  await expectColorToBeSelected(canvas, contentColorsByName.DARK_PURPLE);
+  await clickOnColor(canvas, ContentColorByName.DARK_PURPLE);
+  await expectColorToBeSelected(canvas, ContentColorByName.DARK_PURPLE);
 
   //move with keyboard to a different color
   await typeMultipleTimes("{arrowRight}", 3);
-  await expectColorToBeActive(canvas, contentColorsByName.INDIGO);
+  await expectColorToBeActive(canvas, ContentColorByName.INDIGO);
 
   //select this color as well
   await userEvent.keyboard("{Enter}");
-  await expectColorToBeSelected(canvas, contentColorsByName.DARK_PURPLE);
-  await expectColorToBeSelected(canvas, contentColorsByName.INDIGO);
-  await expectColorToBeActive(canvas, contentColorsByName.INDIGO);
+  await expectColorToBeSelected(canvas, ContentColorByName.DARK_PURPLE);
+  await expectColorToBeSelected(canvas, ContentColorByName.INDIGO);
+  await expectColorToBeActive(canvas, ContentColorByName.INDIGO);
 
   //cancel the selection of the first color
-  await clickOnColor(canvas, contentColorsByName.DARK_PURPLE);
-  await expectColorToBeNotSelected(canvas, contentColorsByName.DARK_PURPLE);
-  await expectColorToBeSelected(canvas, contentColorsByName.INDIGO);
+  await clickOnColor(canvas, ContentColorByName.DARK_PURPLE);
+  await expectColorToBeNotSelected(canvas, ContentColorByName.DARK_PURPLE);
+  await expectColorToBeSelected(canvas, ContentColorByName.INDIGO);
 
   //since we used the mouse, the "active" indicator should be removed
-  await expectColorToBeNotActive(canvas, contentColorsByName.DARK_PURPLE);
-  await expectColorToBeNotActive(canvas, contentColorsByName.INDIGO);
+  await expectColorToBeNotActive(canvas, ContentColorByName.DARK_PURPLE);
+  await expectColorToBeNotActive(canvas, ContentColorByName.INDIGO);
 }
 
 export const multiSelectionInteractionSuite = interactionSuite({