+++ src/components/Dropdown/dropdown-constants.js
@@ -1,7 +1,5 @@
-import Chips from "../Chips/Chips";
-
 export const DROPDOWN_CHIP_COLORS = {
-  PRIMARY: Chips.colors.PRIMARY,
-  NEGATIVE: Chips.colors.NEGATIVE,
-  POSITIVE: Chips.colors.POSITIVE
+  PRIMARY: "PRIMARY",
+  NEGATIVE: "NEGATIVE",
+  POSITIVE: "POSITIVE"
 };