+++ src/utils/colors-vars-map.ts
@@ -174,194 +174,91 @@ export enum ColorStyle {
   SELECTED = "selected"
 }
 
-export const contentColorsByName: Record<string, string> = {
-  GRASS_GREEN: "grass_green",
-  DONE_GREEN: "done-green",
-  BRIGHT_GREEN: "bright-green",
-  SALADISH: "saladish",
-  EGG_YOLK: "egg_yolk",
-  WORKING_ORANGE: "working_orange",
-  DARK_ORANGE: "dark-orange",
-  PEACH: "peach",
-  SUNSET: "sunset",
-  STUCK_RED: "stuck-red",
-  DARK_RED: "dark-red",
-  SOFIA_PINK: "sofia_pink",
-  LIPSTICK: "lipstick",
-  BUBBLE: "bubble",
-  PURPLE: "purple",
-  DARK_PURPLE: "dark_purple",
-  BERRY: "berry",
-  DARK_INDIGO: "dark_indigo",
-  INDIGO: "indigo",
-  NAVY: "navy",
-  BRIGHT_BLUE: "bright-blue",
-  DARK_BLUE: "dark-blue",
-  AQUAMARINE: "aquamarine",
-  CHILI_BLUE: "chili-blue",
-  RIVER: "river",
-  WINTER: "winter",
-  EXPLOSIVE: "explosive",
-  AMERICAN_GRAY: "american_gray",
-  BLACKISH: "blackish",
-  BROWN: "brown",
-  ORCHID: "orchid",
-  TAN: "tan",
-  SKY: "sky",
-  COFFEE: "coffee",
-  ROYAL: "royal",
-  TEAL: "teal",
-  LAVENDER: "lavender",
-  STEEL: "steel",
-  LILAC: "lilac",
-  PECAN: "pecan"
-};
-
-export const stateSelectedColors: Record<string, string> = {
-  POSITIVE: "--positive-color-selected",
-  NEGATIVE: "--negative-color-selected",
-  PRIMARY: "--primary-selected-color"
-};
-export const stateSelectedHoverColors: Record<string, string> = {
-  POSITIVE: "--positive-color-selected-hover",
-  NEGATIVE: "--negative-color-selected-hover",
-  PRIMARY: "--primary-selected-hover-color"
-};
-export const elementAllowedColors = [
-  ...Object.keys(contentColorsByName),
-  ...Object.keys(stateSelectedColors),
-  ...Object.keys(stateSelectedHoverColors)
-];
-export const elementColorsNames = elementAllowedColors.reduce((acc: Record<string, string>, key) => {
-  acc[key] = key;
-  return acc;
-}, {});
-
-export function getElementColor(colorName: string, isSelectedPalette = false, isSelectedHoverPalette = false): string {
-  if (contentColorsByName[colorName]) {
-    return `var(--color-${contentColorsByName[colorName]}${isSelectedPalette ? "-selected" : ""})`;
-  }
-  if (stateSelectedHoverColors[colorName] && isSelectedHoverPalette) {
-    return `var(${stateSelectedHoverColors[colorName]})`;
-  }
-  if (stateSelectedColors[colorName] && isSelectedPalette) {
-    return `var(${stateSelectedColors[colorName]})`;
-  }
-  return colorName;
+export enum ContentColorByName {
+  GRASS_GREEN = "grass_green",
+  DONE_GREEN = "done-green",
+  BRIGHT_GREEN = "bright-green",
+  SALADISH = "saladish",
+  EGG_YOLK = "egg_yolk",
+  WORKING_ORANGE = "working_orange",
+  DARK_ORANGE = "dark-orange",
+  PEACH = "peach",
+  SUNSET = "sunset",
+  STUCK_RED = "stuck-red",
+  DARK_RED = "dark-red",
+  SOFIA_PINK = "sofia_pink",
+  LIPSTICK = "lipstick",
+  BUBBLE = "bubble",
+  PURPLE = "purple",
+  DARK_PURPLE = "dark_purple",
+  BERRY = "berry",
+  DARK_INDIGO = "dark_indigo",
+  INDIGO = "indigo",
+  NAVY = "navy",
+  BRIGHT_BLUE = "bright-blue",
+  DARK_BLUE = "dark-blue",
+  AQUAMARINE = "aquamarine",
+  CHILI_BLUE = "chili-blue",
+  RIVER = "river",
+  WINTER = "winter",
+  EXPLOSIVE = "explosive",
+  AMERICAN_GRAY = "american_gray",
+  BLACKISH = "blackish",
+  BROWN = "brown",
+  ORCHID = "orchid",
+  TAN = "tan",
+  SKY = "sky",
+  COFFEE = "coffee",
+  ROYAL = "royal",
+  TEAL = "teal",
+  LAVENDER = "lavender",
+  STEEL = "steel",
+  LILAC = "lilac",
+  PECAN = "pecan"
 }
 
-export const allMondayColors = [
-  "--color-asphalt",
-  "--color-light_blue",
-  "--color-basic_blue",
-  "--color-basic_light_blue",
-  "--color-dark_blue",
-  "--color-link_color",
-  "--color-snow_white",
-  "--color-success",
-  "--color-success-hover",
-  "--color-success-highlight",
-  "--color-purple",
-  "--color-error",
-  "--color-error-hover",
-  "--color-error-highlight",
-  "--color-placholder_gray",
-  "--color-wolf_gray",
-  "--color-mud_black",
-  "--color-jaco_gray",
-  "--color-black",
-  "--color-dark_purple",
-  "--color-blue_links",
-  "--color-bazooka",
-  "--color-dark_marble",
-  "--color-marble",
-  "--color-gainsboro",
-  "--color-grass_green",
-  "--color-jeans",
-  "--color-egg_yolk",
-  "--color-saladish",
-  "--color-lipstick",
-  "--color-working_orange",
-  "--color-aqua",
-  "--color-brown",
-  "--color-blackish",
-  "--color-explosive",
-  "--color-american_gray",
-  "--color-highlight_blue",
-  "--color-pulse_text_color",
-  "--color-highlight",
-  "--color-placeholder_light_gray",
-  "--color-scrollbar_gray",
-  "--color-timeline_grid_blue",
-  "--color-timeline_blue",
-  "--color-default_group_color",
-  "--color-very_light_gray",
-  "--color-pulse_bg",
-  "--color-jade",
-  "--color-form_purple",
-  "--color-form_btn_hover",
-  "--color-board_views_grey",
-  "--color-board_views_blue",
-  "--color-board_views_grey_hover",
-  "--color-board_views_blue_secondary",
-  "--color-brand-blue",
-  "--color-brand-charcoal",
-  "--color-brand-gold",
-  "--color-brand-green",
-  "--color-brand-iris",
-  "--color-brand-light-blue",
-  "--color-brand-malachite",
-  "--color-brand-purple",
-  "--color-brand-red",
-  "--color-public",
-  "--color-private",
-  "--color-word-blue",
-  "--color-ppt-orange",
-  "--color-excel-green",
-  "--color-pdf-red",
-  "--color-zip-orange",
-  "--color-media-blue",
-  "--color-surface",
-  "--color-burned_eggplant",
-  "--color-live_blue",
-  "--color-extra_light_gray",
-  "--color-glitter",
-  "--color-ultra_light_gray",
-  "--color-red_shadow",
-  "--color-green_shadow",
-  "--color-storm_gray",
-  "--color-riverstone_gray",
-  "--color-ui_grey",
-  "--color-border_light_gray",
-  "--color-like_red",
-  "--color-lime-green",
-  "--color-mustered",
-  "--color-dark_red",
-  "--color-dark-red",
-  "--color-trolley-grey",
-  "--color-dark-purple",
-  "--color-dark-orange",
-  "--color-sofia_pink",
-  "--color-dark-pink",
-  "--color-turquoise",
-  "--color-light-pink",
-  "--color-red-shadow",
-  "--color-orange",
-  "--color-yellow",
-  "--color-green-shadow",
-  "--color-grass-green",
-  "--color-blue-links",
-  "--color-bright-blue",
-  "--color-amethyst",
-  "--color-green-haze",
-  "--color-sunset",
-  "--color-bubble",
-  "--color-peach",
-  "--color-berry",
-  "--color-winter",
-  "--color-river",
-  "--color-navy",
-  "--color-aquamarine",
-  "--color-indigo",
-  "--color-dark_indigo"
-];
+export enum StateSelectedColor {
+  POSITIVE = "--positive-color-selected",
+  NEGATIVE = "--negative-color-selected",
+  PRIMARY = "--primary-selected-color"
+}
+export enum StateSelectedHoverColor {
+  POSITIVE = "--positive-color-selected-hover",
+  NEGATIVE = "--negative-color-selected-hover",
+  PRIMARY = "--primary-selected-hover-color"
+}
+
+export const ElementAllowedColor = {
+  ...ContentColorByName,
+  ...StateSelectedColor,
+  ...StateSelectedHoverColor
+};
+
+export type ElementColor = typeof ElementAllowedColor[keyof typeof ElementAllowedColor] | string;
+
+const getColorKeyByValue = (colorValue: string, colors: Record<string, string>) => {
+  return Object.keys(colors)[Object.values(colors).indexOf(colorValue)];
+};
+
+export function getElementColor(
+  colorValue: keyof typeof ElementAllowedColor | string,
+  isSelectedPalette = false,
+  isSelectedHoverPalette = false
+): string {
+  const colorKey = getColorKeyByValue(colorValue, ElementAllowedColor);
+  if (!colorKey) {
+    return colorValue;
+  }
+
+  if (ContentColorByName[colorKey as keyof typeof ContentColorByName]) {
+    return `var(--color-${ContentColorByName[colorKey as keyof typeof ContentColorByName]}${
+      isSelectedPalette ? "-selected" : ""
+    })`;
+  }
+  if (StateSelectedHoverColor[colorKey as keyof typeof StateSelectedHoverColor] && isSelectedHoverPalette) {
+    return `var(${StateSelectedHoverColor[colorKey as keyof typeof StateSelectedHoverColor]})`;
+  }
+  if (StateSelectedColor[colorKey as keyof typeof StateSelectedColor] && isSelectedPalette) {
+    return `var(${StateSelectedColor[colorKey as keyof typeof StateSelectedColor]})`;
+  }
+}